Add support for groups

+++ telegram_gcloner/handlers/cancel.py
@@ -4,6 +4,8 @@ import logging
 
 from telegram.ext import Dispatcher, CallbackQueryHandler
 
+from utils.restricted import restricted
+
 logger = logging.getLogger(__name__)
 
 
@@ -12,7 +14,11 @@ def init(dispatcher: Dispatcher):
     dispatcher.add_handler(CallbackQueryHandler(cancel, pattern=r'^cancel$'))
 
 
+@restricted
 def cancel(update, context):
     query = update.callback_query
-    # query.message.edit_reply_markup(reply_markup=None)
-    query.message.delete()
+    if query.message.chat_id < 0 and \
+            (not query.message.reply_to_message or
+             query.from_user.id != query.message.reply_to_message.from_user.id):
+        # query.message.edit_reply_markup(reply_markup=None)
+        query.message.delete()
